+++ lib/app/features/home/expense_list/widgets/expenditure_widget.dart
@@ -1,0 +1,92 @@
+import 'package:flutter/material.dart';
+
+class ExpenditureWidget extends StatelessWidget {
+  const ExpenditureWidget({
+    super.key,
+    required this.width,
+    required this.height,
+    required this.doc,
+  });
+
+  final double width;
+  final double height;
+  final String doc;
+  // final concreteExpense = TextEditingController();
+  // final cost = TextEditingController();
+
+  @override
+  Widget build(BuildContext context) {
+    return InkWell(
+      onTap: () {},
+      child: Container(
+        margin: const EdgeInsets.all(10),
+        decoration: BoxDecoration(
+            borderRadius: const BorderRadius.only(
+              topLeft: Radius.circular(30),
+              bottomLeft: Radius.circular(30),
+              topRight: Radius.circular(30),
+              bottomRight: Radius.circular(15),
+            ),
+            border: Border.all(color: Colors.orange)),
+        child: Column(
+          children: [
+            Row(
+              mainAxisAlignment: MainAxisAlignment.spaceBetween,
+              children: [
+                Container(
+                  width: width * 0.12,
+                  height: height * 0.06,
+                  decoration: BoxDecoration(
+                    color: Colors.orange,
+                    borderRadius: BorderRadius.circular(30),
+                  ),
+                ),
+                SizedBox(
+                  width: width * 0.5,
+                  height: height * 0.06,
+                  child: Column(
+                    mainAxisAlignment: MainAxisAlignment.center,
+                    children: [
+                      Text(
+                        doc,
+                        style:
+                            const TextStyle(fontSize: 16, color: Colors.orange),
+                        textAlign: TextAlign.center,
+                      )
+                    ],
+                  ),
+                ),
+                Container(
+                  width: width * 0.20,
+                  height: height * 0.06,
+                  decoration: const BoxDecoration(
+                    color: Colors.orange,
+                    borderRadius: BorderRadius.only(
+                      topLeft: Radius.circular(10),
+                      bottomLeft: Radius.circular(20),
+                      topRight: Radius.circular(20),
+                      bottomRight: Radius.circular(10),
+                    ),
+                  ),
+                  child: Column(
+                    mainAxisAlignment: MainAxisAlignment.center,
+                    children: const [
+                      Padding(
+                        padding: EdgeInsets.all(2.0),
+                        child: Text(
+                          '0 PLN',
+                          style: TextStyle(color: Colors.white, fontSize: 16),
+                          textAlign: TextAlign.center,
+                        ),
+                      ),
+                    ],
+                  ),
+                )
+              ],
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
