+++ lib/app/features/home/my_account/cubit/my_account_state.dart
@@ -1,13 +1,13 @@
 part of 'my_account_cubit.dart';
 
 class MyAccountState {
-  final List<QueryDocumentSnapshot<Map<String, dynamic>>> documents;
+  final List<QueryDocumentSnapshot<Map<String, dynamic>>>? documents;
   final String errorMessage;
   final bool loading;
 
   MyAccountState({
-    required this.documents,
-    required this.errorMessage,
-    required this.loading,
+    this.documents,
+    this.errorMessage = '',
+    this.loading = false,
   });
 }
